# src/pkg/server/handler/apihandler/embedded/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SMS Service API",
    "version": "1.0.0",
    "license": {
      "name": "unlicensed",
      "url": "www.example.com"
    },
    "contact": {
      "name": "API Support",
      "email": "support@example.com",
      "url": "http://example.com"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "This server"
    }
  ],
...
  "paths": {
...
    "/sms/basic/send": {
      "post": {
        "description": "Send a SMS with a custom message text by giving the recipients data as well as the message text",
        "summary": "Send a SMS with custom text",
        "operationId": "sendBasicSMS",
        "tags": [
          "SMS"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SMSSendRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "SMS send"
          },
          "400": {
            "description": "Invalid receiver phone number supplied"
          }
        },
        "security": []
      }
    },
    "/sms/template/{templateName}/send": {
      "post": {
        "description": "Send a templated SMS with a custom message text by giving the recipients data as well as the placeholder values",
        "summary": "Send a templated SMS with custom text",
        "operationId": "sendTemplatedSMS",
        "tags": [
          "SMS"
        ],
        "parameters": [
          {
            "name": "templateName",
            "in": "path",
            "description": "Name of template to fill and send",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "exampleTemplate"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SMSTemplateSendRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "SMS send"
          },
          "400": {
            "description": "Invalid receiver phone number supplied"
          }
        },
        "security": []
      }
    },
    "/sms/templates/{templateName}": {
      "get": {
        "description": "Get a specific SMS template by its name",
        "summary": "Get SMS Template by Name",
        "operationId": "getSMSTemplateByName",
        "tags": [
          "SMS"
        ],
        "parameters": [
          {
            "name": "templateName",
            "in": "path",
            "description": "Name of template to return",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "exampleTemplate"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SMSTemplate"
                }
              }
            }
          },
          "400": {
            "description": "Invalid name supplied"
          },
          "404": {
            "description": "Can not find template (template name)"
          }
        },
        "security": []
      },
      "post": {
        "description": "Add a new SMS template to the db",
        "summary": "Add new SMS template",
        "operationId": "addNewSMSTemplate",
        "tags": [
          "SMS"
        ],
        "parameters": [
          {
            "name": "templateName",
            "in": "path",
            "description": "Name of template to add",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "exampleTemplate"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SMSTemplatePostRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Added template (name)"
          },
          "400": {
            "description": "Invalid name supplied"
          }
        },
        "security": []
      }
    },
    "/sms/templates/{templateName}/placeholders": {
      "get": {
        "description": "Get all placeholders of a specific SMS template that have to be filled with some data",
        "summary": "Get SMS Template Placeholders",
        "operationId": "getSMSTemplatePlaceholdersByName",
        "tags": [
          "SMS"
        ],
        "parameters": [
          {
            "name": "templateName",
            "in": "path",
            "description": "Name of SMS template to return",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SMSPlaceholders"
                }
              }
            }
          },
          "400": {
            "description": "Invalid name supplied"
          },
          "404": {
            "description": "Can not find template (template name)"
          }
        },
        "security": []
      }
    },
    "/sms/templates/{templateName}/fill": {
      "post": {
        "description": "Fill the placeholders of a specific SMS template.",
        "summary": "Fill placeholders of SMS template",
        "operationId": "fillSMSTemplate",
        "tags": [
          "SMS"
        ],
        "parameters": [
          {
            "name": "templateName",
            "in": "path",
            "description": "Name of template to fill",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SMSTemplateSendRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SMSFilledTemplateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid name or data supplied"
          }
        },
        "security": []
      }
    },
...
    "/workflow/{workflowName}": {
      "post": {
        "description": "Create a new workflow for sending emails with or without static and templated attachments",
        "summary": "Create a new workflow",
        "operationId": "createWorkflow",
        "tags": [
          "Workflow"
        ],
        "parameters": [
          {
            "name": "workflowName",
            "in": "path",
            "description": "Name of workflow to return",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "exampleWorkflow"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workflow created"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "security": []
      },
      "get": {
        "description": "Get a specific workflow by its name",
        "summary": "Get Workflow by Name",
        "operationId": "getWorkflowByName",
        "tags": [
          "Workflow"
        ],
        "parameters": [
          {
            "name": "workflowName",
            "in": "path",
            "description": "Name of workflow to return",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "exampleWorkflow"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetWorkflowResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid name supplied"
          },
          "404": {
            "description": "Can not find workflow (workflow name)"
          }
        },
        "security": []
      }
    },
    "/workflow/{workflowName}/send": {
      "post": {
        "description": "Use a workflow to send emails with or without static and templated attachments",
        "summary": "Use workflow",
        "operationId": "useWorkflow",
        "tags": [
          "Workflow"
        ],
        "parameters": [
          {
            "name": "workflowName",
            "in": "path",
            "description": "Name of workflow to use",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "exampleWorkflow"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowSendRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workflow used"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "security": []
      }
    }
  },
  "components": {
    "schemas": {
...
      "SMSSendRequest": {
        "type": "object",
        "properties": {
          "receiverPhoneNumber": {
            "type": "string",
            "example": "+4915170640522"
          },
          "message": {
            "type": "string",
            "example": "Don't forget to buy some milk!"
          }
        },
        "required": [
          "receiverPhoneNumber",
          "message"
        ]
      },
      "Placeholder": {
        "type": "object",
        "additionalProperties": {
          "type": "string"
        },
        "example": {
          "Name": "Sherlock Holmes",
          "Age": "45"
        }
      },
      "SMSTemplateSendRequest": {
        "type": "object",
        "properties": {
          "receiverPhoneNumber": {
            "type": "string",
            "example": "+4915170640522"
          },
          "placeholders": {
            "$ref": "#/components/schemas/Placeholder"
          }
        },
        "required": [
          "receiverPhoneNumber",
          "placeholders"
        ]
      },
      "SMSTemplate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "ExampleName"
          },
          "templateString": {
            "type": "string",
            "example": "Hello {{.Name}}. You are {{.Age}} years old."
          }
        },
        "required": [
          "name",
          "templateString"
        ]
      },
      "SMSTemplatePostRequest": {
        "type": "object",
        "properties": {
          "templateString": {
            "type": "string",
            "example": "Hello {{.Name}}. You are {{.Age}} years old."
          }
        },
        "required": [
          "templateString"
        ]
      },
      "SMSPlaceholders": {
        "type": "object",
        "properties": {
          "placeholders": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "key",
                      "value"
                    ],
                    "properties": {
                      "key": {
                        "type": "string"
                      },
                      "value": {
                        "type": "string"
                      }
                    },
                    "example": [
                      {
                        "key": "firstName",
                        "value": ""
                      },
                      {
                        "key": "lastName",
                        "value": ""
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "required": [
          "placeholders"
        ],
        "example": {
          "placeholders": [
            {
              "data": [
                {
                  "key": "firstName",
                  "value": ""
                },
                {
                  "key": "lastName",
                  "value": ""
                }
              ]
            }
          ]
        }
      },
      "SMSFilledTemplateResponse": {
        "type": "string",
        "example": "Hello Sherlock Holmes, you are 45 years old."
      },
...
      "TemplateInfo": {
        "type": "object",
        "properties": {
          "templateName": {
            "type": "string",
            "example": "exampleName"
          },
          "placeholders": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "examplePlaceholder1",
              "examplePlaceholder2"
            ]
          }
        },
        "required": [
          "templateName",
          "placeholders"
        ]
      },
      "GetWorkflowResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "ExampleName"
          },
          "emailSubject": {
            "type": "string",
            "example": "Example Subject"
          },
          "emailTemplate": {
            "$ref": "#/components/schemas/TemplateInfo"
          },
          "pdfTemplates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemplateInfo"
            }
          },
          "staticAttachments": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "exampleAttachment"
            }
          }
        },
        "required": [
          "name",
          "emailSubject",
          "emailTemplate",
          "pdfTemplates",
          "staticAttachments"
        ]
      },
      "WorkflowCreateRequest": {
        "type": "object",
        "properties": {
          "emailTemplateName": {
            "type": "string",
            "example": "exampleTemplate"
          },
          "staticAttachmentNames": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "example.pdf"
            }
          },
          "templatedAttachmentNames": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "exampleTemplate"
            }
          },
          "emailSubject": {
            "type": "string",
            "example": "Registration Confirmation"
          }
        },
        "required": [
          "emailTemplateName",
          "staticAttachmentNames",
          "templatedAttachmentNames",
          "emailSubject"
        ]
      },
      "FillTemplate": {
        "type": "object",
        "description": "A request to fill a template with placeholders.\n",
        "properties": {
          "templateName": {
            "type": "string",
            "example": "exampleTemplate"
          },
          "placeholders": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "example": {
              "Name": "Sherlock Holmes",
              "Age": "45"
            }
          }
        }
      },
      "WorkflowSendRequest": {
        "type": "object",
        "properties": {
          "toEmail": {
            "type": "string",
            "example": "example@example.com"
          },
          "toName": {
            "type": "string",
            "example": "exampleName"
          },
          "emailTemplate": {
            "$ref": "#/components/schemas/FillTemplate"
          },
          "pdfTemplates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FillTemplate"
            }
          }
        },
        "required": [
          "toEmail",
          "toName",
          "emailTemplate"
        ]
      },
...
    }
  }
}